Create groove-ml-2.py
A second model in progress with varied features now including the onset envelope with a much lower hop length to detect onset strengths with much more accuracy. The model also includes a self recurrence matrix constructed from the spectral flatness (a measure of how noise-like the audio frame is) which shows a lot of promise. Also experimenting with the smoothing/use of diagonal filters with the onset envelope's self recurrence matrix. Much higher hop length must be used here (given exponential calculations required) so unsure of usefulness as of now.